fix: game details showing live when its in unofficial final
The api returns a "final" field that is a string of either "1" or "0" if
the game is over and its finalized. However, when the details have a
"status" field of "Unofficial Final" the "final" field is "0" (meaning
it's still in progress). To fix this, a game is considered final when
either "final" = "1" OR "status" = "Unofficial Final". This way game
details stop showing the game as live when they aren't actually live
anymore.